rgbdmux: don't error out on failure pushing Gap event
gst_pad_push_event() might return false when downstream starts
flushing, this cannot be cause for erroring out altogether.

The usual workaround is to log a warning, ignore the error and
potentially let an actual, ulterior call to gst_pad_push() trigger
an error when more information is available in the return value.